Switch use of google-flags (gflags) to more standard argparse.
It is more appropriate for open source projects to use argparse.

Keep the 'clever' approach and uses the docstring on methods on
AdbCommands/FastbootCommands to generate the arguments on the parser. Override
the documentation only when it doesn't make sense, like when an argument can be
"object-like".

Update default fastboot packet size from 4kb to 1Mb.

Fix tests (adb_test.py was broken). Make them executable.